requirements.txt
arrow==0.17.0
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.2.0
autopep8==1.5.4
Babel==2.8.0
blessed==1.17.11
boto3==1.16.23
botocore==1.19.23
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.3
django-3-jet==1.0.8
django-admin-sortable2==0.7.7
django-cors-headers==3.5.0
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-picklefield==3.0.1
django-q==1.3.4
django-storages==1.10.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-nested-routers==0.92.1
drf-yasg==1.20.0
drf-yasg2==1.19.4
et-xmlfile==1.0.1
Faker==4.17.1
firebase-admin==4.4.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==2.0.1
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
gprof2dot==2019.11.30
grpcio==1.33.2
gunicorn==20.1.0
httplib2==0.18.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
msgpack==1.0.0
multidict==4.7.6
openpyxl==3.0.5
packaging==20.4
phonenumberslite==8.12.13
Pillow==8.0.1
proto-plus==1.11.0
protobuf==3.14.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest-django==4.1.0
pytest-factoryboy==2.0.3
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
redis==3.5.3
requests==2.25.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
sentry-sdk==0.16.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.1
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
yarl==1.6.0
django-debug-toolbar==3.2
django-admin-rangefilter==0.6.3
django-elasticsearch-dsl==7.1.4
django-queryset-csv==1.1.0
python-dotenv==0.17.1
whitenoise==5.2.0
json2xml==3.6.0
redis==3.5.3